+++ storedimo/database/migrations/2025_03_25_073150_create_compra_productos.php
@@ -13,20 +13,30 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('compra_productos', function (Blueprint $table) {
-            $table->increments('id_compra_producto');
-            $table->unsignedInteger('id_compra')->nullable();
-            $table->unsignedInteger('id_producto')->nullable();
-            $table->string('cantidad')->nullable();
-            $table->string('precio_unitario_compra')->nullable();
-            $table->string('subtotal')->nullable();
+        if (!Schema::hasTable('compra_productos'))
+        {
+            Schema::create('compra_productos', function (Blueprint $table) {
+                $table->increments('id_compra_producto');
+                $table->unsignedInteger('id_compra')->nullable();
+                $table->unsignedInteger('id_producto')->nullable();
+                $table->string('cantidad')->nullable();
+                $table->string('precio_unitario_compra')->nullable();
+                $table->string('subtotal')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->timestamps();
-            $table->softDeletes();
+                if (Schema::hasTable('compras'))
+                {
+                    $table->foreign('id_compra')->references('id_compra')->on('compras')->onDelete('cascade');
+                }
 
-            $table->foreign('id_compra')->references('id_compra')->on('compras')->onDelete('cascade');
-            $table->foreign('id_producto')->references('id_producto')->on('productos')->onDelete('cascade');
-        });
+                if (Schema::hasTable('productos'))
+                {
+                    $table->foreign('id_producto')->references('id_producto')->on('productos')->onDelete('cascade');
+                }
+            });
+
+        }
     }
 
     /**
@@ -36,6 +46,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('compra_productos');
+        if (Schema::hasTable('compra_productos'))
+        {
+            Schema::dropIfExists('compra_productos');
+        }
     }
 };
